refactor: helper constants placement
