+++ src/commands/navigation/cd.js
@@ -1,4 +1,4 @@
-import { COMMAND_SEPARATOR, isCommand } from '../../helpers/common.js'
+import { COMMAND_SEPARATOR, isCommand } from '../../helpers/cli.js'
 import { isDirectory } from '../../helpers/fs.js'
 import { store } from '../../store.js'
 import path from 'node:path'
